Refactor Destination model: rename fields for consistency and add conversion method; update SupabaseAgencyRepository to use new SupabaseDestination class.
