--- app/src/main/java/com/android/tripbook/model/Destination.kt
@@ -1,9 +1,0 @@
-package com.android.tripbook.model
-
-data class Destination(
-    val id: Int,
-    val agencyId: Int,
-    val agencyRating: Int,
-    val destinationName: String,
-    val destinationTarif: Double
-)
