+++ app/src/main/java/com/android/tripbook/repository/SupabaseAgencyRepository.kt
@@ -3,7 +3,6 @@ package com.android.tripbook.repository
 import android.util.Log
 import com.android.tripbook.data.SupabaseConfig
 import com.android.tripbook.model.Agency
-import com.android.tripbook.model.Destination
 import io.github.jan.supabase.postgrest.from
 import io.github.jan.supabase.exceptions.RestException
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -35,25 +34,13 @@ data class SupabaseAgency(
 }
 
 @Serializable
-data class SupabaseDestination(
+data class Destination(
     val id: Int,
-    val agencyid: Int,
-    val agency_rating: Int,
-    val destination_name: String,
-    val destination_tarif: Double
-) {
-    fun toDestination(): Destination {
-        return Destination(
-            id = id,
-            agencyId = agencyid,
-            agencyRating = agency_rating,
-            destinationName = destination_name,
-            destinationTarif = destination_tarif
-        )
-    }
-}
-
-
+    val agencyId: Int,
+    val agencyRating: Int,
+    val destinationName: String,
+    val destinationTarif: Double
+)
 
 class SupabaseAgencyRepository {
     private val supabase = SupabaseConfig.client
@@ -104,16 +91,13 @@ class SupabaseAgencyRepository {
             Log.d(TAG, "Loading destinations for agency $agencyId...")
 
             // Load all destinations first, then filter in-memory
-            val allSupabaseDestinations = supabase.from(DESTINATIONS_TABLE)
+            val allDestinations = supabase.from(DESTINATIONS_TABLE)
                 .select()
-                .decodeList<SupabaseDestination>()
+                .decodeList<Destination>()
 
-            Log.d(TAG, "Loaded ${allSupabaseDestinations.size} destinations from database")
-            Log.d(TAG, "Raw Supabase destinations: $allSupabaseDestinations")
+            Log.d(TAG, "Loaded ${allDestinations.size} destinations from database: $allDestinations")
 
-            val filteredDestinations = allSupabaseDestinations
-                .map { it.toDestination() }
-                .filter { it.agencyId == agencyId }
+            val filteredDestinations = allDestinations.filter { it.agencyId == agencyId }
                 .sortedBy { it.destinationName }
 
             Log.d(TAG, "Filtered ${filteredDestinations.size} destinations for agency $agencyId: $filteredDestinations")
@@ -121,8 +105,6 @@ class SupabaseAgencyRepository {
             filteredDestinations
         } catch (e: Exception) {
             Log.e(TAG, "Error loading destinations: ${e.message}", e)
-            Log.e(TAG, "Exception type: ${e.javaClass.simpleName}")
-            e.printStackTrace()
             _error.value = "Failed to load destinations: ${e.message}"
             emptyList()
         } finally {
